Rename certificates as navigationEvents
