+++ src/storeAndEffects.js
@@ -13,8 +13,8 @@ import { isUpdatingEnabled } from './reducers/isUpdatingEnabled';
 import { servers } from './reducers/servers';
 import { skippedUpdateVersion } from './reducers/skippedUpdateVersion';
 import { spellCheckingDictionaries } from './reducers/spellCheckingDictionaries';
-import { certificatesSaga } from './sagas/certificates';
 import { deepLinksSaga } from './sagas/deepLinks';
+import { navigationEventsSaga } from './sagas/navigationEvents';
 import { serversSaga } from './sagas/servers';
 import { spellCheckingSaga } from './sagas/spellChecking';
 import { updatesSaga } from './sagas/updates';
@@ -34,8 +34,8 @@ const rootReducer = combineReducers({
 });
 
 function *rootSaga() {
-	yield fork(certificatesSaga);
 	yield fork(deepLinksSaga);
+	yield fork(navigationEventsSaga);
 	yield fork(serversSaga);
 	yield fork(spellCheckingSaga);
 	yield fork(updatesSaga);
